Extracted common model Constants
