+++ LudoCompose/app/src/main/java/hu/bme/aut/android/ludocompose/ui/model/Constants.kt
@@ -17,6 +17,9 @@
 package hu.bme.aut.android.ludocompose.ui.model
 
 import androidx.compose.ui.geometry.Offset
+import hu.bme.aut.android.ludocompose.common.model.Constants.pieceCount
+import hu.bme.aut.android.ludocompose.common.model.Constants.playerMaxCount
+import hu.bme.aut.android.ludocompose.common.model.Constants.trackSize
 import kotlin.math.PI
 import kotlin.math.cos
 import kotlin.math.sin
@@ -27,11 +30,6 @@ data object Constants {
     private val Int.rad get() = (mod(360) * RAD).toFloat()
     private fun cos(deg: Int) = cos(deg.rad)
     private fun sin(deg: Int) = sin(deg.rad)
-
-    const val playerCount = 4
-    const val pieceCount = 4
-    const val trackMultiplier = 10
-    const val trackSize = trackMultiplier * playerCount
 
     // calculate points in a 1x1 clip square
     const val width = 1f
@@ -50,7 +48,7 @@ data object Constants {
 
     val diceOffset = Offset(halfWidth, halfHeight)
 
-    private const val playerAngleStep = 360 / playerCount
+    private const val playerAngleStep = 360 / playerMaxCount
     private const val tokenAngleStep = 360 / pieceCount
     private const val trackAngleStep = 360 / trackSize
 
@@ -58,7 +56,7 @@ data object Constants {
     private const val trackStartAngle = startAngle + trackAngleStep
 
     val yardPoints by lazy {
-        Array(playerCount) { player ->
+        Array(playerMaxCount) { player ->
             val outerAngle = player * playerAngleStep + startAngle
             val outerRadiusX = halfWidth + trackPadding
             val outerRadiusY = halfHeight + trackPadding
@@ -89,7 +87,7 @@ data object Constants {
     }
 
     val homePoints by lazy {
-        Array(playerCount) { player ->
+        Array(playerMaxCount) { player ->
             val angle = player * playerAngleStep + startAngle
             Array(pieceCount) { piece ->
                 val radiusDiff = (piece + 2) * trackPadding
